+++ docs/CHANGES.rst
@@ -12,6 +12,8 @@ Change Log
   * ``nemo_cmd.prepare.check_nemo_exec()``
   * ``nemo_cmd.prepare.check_xios_exec()``
   * ``nemo_cmd.prepare.copy_run_set_files()``
+  * ``nemo_cmd.prepare.get_hg_revision()``
+  * ``nemo_cmd.prepare.get_n_processors()``
   * ``nemo_cmd.prepare.make_executable_links()``
   * ``nemo_cmd.prepare.make_forcing_links()``
   * ``nemo_cmd.prepare.make_grid_links()``